requirements.txt
discord.py==2.3.2
requests==2.32.2
flask
python-dotenv
pytz
beautifulsoup4